Added separate instance for MonadError
